requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
Deprecated==1.2.13
exceptiongroup==1.0.4
firebase-admin==6.0.1
Flask==2.2.2
gcloud==0.18.3
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
msgpack==1.0.4
oauth2client==4.1.3
outcome==1.2.0
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase4==4.5.0
PySocks==1.7.1
python-jwt==4.0.0
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
